Get shill ready for slotted package builds.
This is the subset of https://gerrit.chromium.org/gerrit/16156
that will allow shill to build once the slotting mechanism of
https://gerrit.chromium.org/gerrit/15415 is in place. It fixes
some warnings that are the result of using slotting at all,
rather than any particular version of libchrome. This will
decouple the submission of the slotting CL from the changes
to upgrade shill to the new callback mechanism.

BUG=chromium-os:16623
TEST=Build shill and unit tests.

Reviewed-on: https://gerrit.chromium.org/gerrit/17253
Tested-by: Eric Shienbrood <ers@chromium.org>
Commit-Ready: Eric Shienbrood <ers@chromium.org>

+++ shill/dns_client.cc
@@ -198,7 +198,8 @@ void DNSClient::ReceiveDNSReply(int status, struct hostent *hostent) {
   if (status == ARES_SUCCESS &&
       hostent != NULL &&
       hostent->h_addrtype == address_.family() &&
-      hostent->h_length == IPAddress::GetAddressLength(address_.family()) &&
+      static_cast<size_t>(hostent->h_length) ==
+      IPAddress::GetAddressLength(address_.family()) &&
       hostent->h_addr_list != NULL &&
       hostent->h_addr_list[0] != NULL) {
     address_ = IPAddress(address_.family(),
